api_gateway/pyproject.toml
[project]
name = "api_gateway"
version = "1.0.0"
description = ""
authors = [
    {name = "Daniil Astapenko", email = "xastapen@mendelu.cz"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.26.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.3",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "python-multipart>=0.0.6",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload --port 8000"
grpc = "python -m grpc_tools.protoc -I../protos --python_out=./src/grpc_output --pyi_out=./src/grpc_output --grpc_python_out=./src/grpc_output ../protos/itinerary.proto"